backend/requirements.txt
SQLAlchemy==2.0.44
psycopg2-binary
python-dotenv
fastapi==0.104.1
uvicorn[standard]
pydantic==2.8.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
openfoodfacts